all working pre test

+++ printf/helpers/helpers.h
@@ -6,7 +6,7 @@
 /*   By: rburri <rburri@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 09:21:58 by rburri            #+#    #+#             */
-/*   Updated: 2021/11/09 09:13:10 by rburri           ###   ########.fr       */
+/*   Updated: 2021/11/09 09:53:52 by rburri           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int		ft_putchar(char c);
 int		ft_print_c(char c);
 int		ft_print_u(unsigned int num);
 int		ft_print_s(char *str);
-int		ft_print_p(unsigned int num);
+int		ft_print_p(size_t num);
 int		ft_print_int(int num);
 int		ft_print_cx(unsigned int num);
 int		ft_print_x(unsigned int num);
